:bug: fix: routes correction
