+++ front/src/App.jsx
@@ -1,8 +1,9 @@
 import { Routes, Route } from 'react-router-dom';
 import Header from './components/Header';
 import LoginPage from './pages/LoginPage';
-import HomePage from './pages/HomePage';
-import UsersPage from './pages/SignupPage'; 
+import HomePage from './pages/InitialPage';
+import SignupPage from './pages/SignupPage';
+import ProtectedHomePage from './pages/HomePage'; 
 
 function App() {
   return (
@@ -11,8 +12,8 @@ function App() {
       <Routes>
         <Route path="/" element={<HomePage />} />
         <Route path="/login" element={<LoginPage />} />
-        <Route path="/users" element={<UsersPage />} />
-        <Route path="/home" element={<UsersPage />} />
+        <Route path="/signup" element={<SignupPage />} />
+        <Route path="/home" element={<ProtectedHomePage />} />
       </Routes>
     </div>
   );
