+++ front/src/pages/HomePage.jsx
@@ -1,18 +1,12 @@
-import { Link } from 'react-router-dom';
 import './PageStyles.css';
-import lockIcon from '../assets/lock-icon.png'; 
 
-function HomePage() {
+function ProtectedHomePage() {
   return (
     <div className="page">
-      <img src={lockIcon} alt="Lock Icon" className="lock-icon" />
-      <h1>Eletronic Lock App</h1>
-      <p>Controle sua fechadura eletrônica de forma prática e segura!</p>
-      <Link to="/login" className="page-button">
-        Sign in
-      </Link>
+      <h1>Acesso liberado!!</h1>
+      <p>Você está logado.</p>
     </div>
   );
 }
 
-export default HomePage;
+export default ProtectedHomePage;
